+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Changelog for NeoFS Node
 - `node` config option `storage.ignore_uninited_shards` (#2953)
 - For `neofs-cli container create`, add `--global-name` flag, that sets name attribute as the value of `__NEOFS__NAME` 
 attribute, which is used for container domain name in NNS contracts (#2954)
+- Save last epoch when metabase was resynchronized (#2966)
+- `neofs-lens meta last-resync-epoch` command (#2966)
 
 ### Fixed
 - Do not search for tombstones when handling their expiration, use local indexes instead (#2929)